// examples/start-full-stack/src/generated/api/schema.ts
export const userRoleSchema = z.enum(["admin", "user", "guest"])
export const userSchema = z.object({
  id: z.string(),
  email: z.email(),
  name: z.string(),
  role: userRoleSchema,
  createdAt: z.iso.datetime(),
  updatedAt: z.iso.datetime()
})
export const createUserInputSchema = z.object({
  email: z.email(),
  name: z.string(),
  role: userRoleSchema
})
export const updateUserInputSchema = z.object({
  email: z.email().nullish(),
  name: z.string().nullish(),
  role: userRoleSchema.nullish()
})
// ...
export const updateUserRequestSchema = updateUserInputSchema
// ...
export type UserRole = z.infer<typeof userRoleSchema>
// ...
export type UpdateUserRequest = z.infer<typeof updateUserRequestSchema>